chunk faceculling, updated lightning, voxel copy/paste mechanics
- Fixed faces between chunks:
While generating the world, the greedymesher takes a look into neighboring chunks for any processed border voxel. The mesher checks if each of these voxels contains the same material type as their neighbor from a different chunk. If this is true the face in between is not added to the mesh any longer.

- Fixed lightning:
The time referred by the renderer is now 100% equal to the real-world time. This means, at 6am in reallife the sun will rise ingame, and at 6pm in reallife the night will begin ingame. However this mechanic can be edited by changing parameters. At the moment an ingame hour is mapped to a reallife hour by 3600 seconds. By changing the numbber of seconds per ingame hour the ingame daylength can be increased or decreased.

- Copy / paste voxels:
Added a mechanic to copy or paste voxels within a user-specified volume.

- Coding style:
Updated coding style